refractor: Break ProductDetails component into smaller components
